requirements.txt
groq>=0.4.1
python-docx>=0.8.11
fpdf2>=2.7.0
PyPDF2>=3.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
